[refactor - #52] 커플 정보 조회 CoupleReadService로 이동
