+++ src/main/java/com/dateplan/dateplan/domain/couple/controller/CoupleController.java
@@ -4,6 +4,7 @@ import com.dateplan.dateplan.domain.couple.dto.CoupleInfoResponse;
 import com.dateplan.dateplan.domain.couple.dto.CoupleInfoServiceResponse;
 import com.dateplan.dateplan.domain.couple.dto.FirstDateRequest;
 import com.dateplan.dateplan.domain.couple.dto.FirstDateResponse;
+import com.dateplan.dateplan.domain.couple.service.CoupleReadService;
 import com.dateplan.dateplan.domain.couple.service.CoupleService;
 import com.dateplan.dateplan.global.dto.response.ApiResponse;
 import jakarta.validation.Valid;
@@ -21,6 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class CoupleController {
 
 	private final CoupleService coupleService;
+	private final CoupleReadService coupleReadService;
 
 	@Deprecated
 	@GetMapping("/{couple_id}/first-date")
@@ -38,7 +40,7 @@ public class CoupleController {
 
 	@GetMapping("/me")
 	public ApiResponse<CoupleInfoResponse> getCoupleInfo() {
-		CoupleInfoServiceResponse response = coupleService.getCoupleInfo();
+		CoupleInfoServiceResponse response = coupleReadService.getCoupleInfo();
 		return ApiResponse.ofSuccess(response.toCoupleInfoResponse());
 	}
 }
